+++ src/components/Login/Login.jsx
@@ -1,22 +1,64 @@
-import { useRef } from "react";
+import { useContext, useRef } from "react";
 import { $Login, $Form, $Input, $Button, $Logo } from "./Login.styles.jsx";
 import logo from "../../assets/carbon-tracker-logo-pure.svg";
+import { EmployeeContext } from "../../store/employee-context.jsx";
+import { apiCallForAuthentication } from "../../api/Api.jsx";
 
-const Login = () => {  
-    
-    const usernameRef = useRef(null); 
-    const passwordRef = useRef(null);
+const Login = () => {
+  const { setEmployeeData } = useContext(EmployeeContext);
 
-    return (
-        <$Login>
-            <$Logo src={logo} alt="Carbon Tracker"/>
-            <$Form>
-                <$Input type="text" ref={usernameRef} placeholder="Corporate key" required/>
-                <$Input type="password" ref={passwordRef} placeholder="Password" required/>
-                <$Button type="submit">Login</$Button>
-            </$Form>
-        </$Login>
+  const usernameRef = useRef(null);
+  const passwordRef = useRef(null);
+
+  const handleLogin = async (event) => {
+    event.preventDefault();
+    setUserCredentials();
+    const employeeData = await apiCallForAuthentication(
+      usernameRef.current.value
+    ).catch((error) => {
+      throw error;
+    });
+    setEmployeeData((_prevData) => ({ ...employeeData }));
+    setUserData({...employeeData})
+  };
+
+  function setUserCredentials() {
+    sessionStorage.setItem(
+        "userCredentials",
+        JSON.stringify({
+          username: usernameRef.current.value,
+          password: passwordRef.current.value,
+        })
+      );
+  }
+
+  function setUserData(userData) {
+    sessionStorage.setItem(
+        "userData",
+        JSON.stringify(userData) 
     );
+  }
+
+  return (
+    <$Login onSubmit={handleLogin}>
+      <$Logo src={logo} alt="Carbon Tracker" />
+      <$Form>
+        <$Input
+          type="text"
+          ref={usernameRef}
+          placeholder="Corporate key"
+          required
+        />
+        <$Input
+          type="password"
+          ref={passwordRef}
+          placeholder="Password"
+          required
+        />
+        <$Button type="submit">Login</$Button>
+      </$Form>
+    </$Login>
+  );
 };
 
 export default Login;